backend/pocket: refactor sign address function
PocketWidgetSignAddress was not only dedicated to the Pocker workflow
anymore. This refactors the function to a more generic `SignBTCAddress`
and moves it into the btc package.
