+++ backend/coins/btc/account.go
@@ -16,6 +16,7 @@
 package btc
 
 import (
+	"encoding/base64"
 	"fmt"
 	"os"
 	"path"
@@ -39,6 +40,7 @@ import (
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc/types"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/coin"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/ltc"
+	"github.com/digitalbitbox/bitbox-wallet-app/backend/exchanges"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/signing"
 	"github.com/digitalbitbox/bitbox-wallet-app/util/errp"
 	"github.com/digitalbitbox/bitbox-wallet-app/util/locker"
@@ -828,4 +830,76 @@ func (account *Account) VerifyExtendedPublicKey(signingConfigIndex int) (bool, e
 		)
 	}
 	return false, nil
+}
+
+// requestAddressScriptTypeMap maps from format codes specified by request-address js library
+// to our own script type codes. See https://www.npmjs.com/package/request-address
+var requestAddressScriptTypeMap = map[string]signing.ScriptType{
+	"p2pkh":  signing.ScriptTypeP2PKH,
+	"p2wpkh": signing.ScriptTypeP2WPKH,
+	"p2sh":   signing.ScriptTypeP2WPKHP2SH,
+	"p2tr":   signing.ScriptTypeP2TR,
+}
+
+// defaultScriptType is the default type used for address sharing and verification, if the
+// `format` param is not defined.
+var defaultScriptType = "p2wpkh"
+
+// SignBTCAddress returns an unused address and makes the user sign a message to prove ownership.
+// Input params:
+//
+//	`account` is the account from which the address is derived, and that will be linked to the Pocket order.
+//	`message` is the message that will be signed by the user with the private key linked to the address.
+//	`format` is the script type that should be used in the address derivation, as received by the widget
+//		(see https://github.com/pocketbitcoin/request-address#requestaddressv0messagescripttype).
+//		If format is empty, native segwit type is used as a fallback.
+//
+// Returned values:
+//
+//	#1: is the first unused address corresponding to the account and the script type identified by the input values.
+//	#2: base64 encoding of the message signature, obtained using the private key linked to the address.
+//	#3: is an optional error that could be generated during the execution of the function.
+func SignBTCAddress(account accounts.Interface, message string, format string) (string, string, error) {
+	if !exchanges.IsPocketSupported(account) {
+		return "", "", errp.Newf("Coin not supported %s", account.Coin().Code())
+	}
+
+	keystore, err := account.Config().ConnectKeystore()
+	if err != nil {
+		return "", "", err
+	}
+
+	canSign := keystore.CanSignMessage(account.Coin().Code())
+	if !canSign {
+		return "", "", errp.Newf("The connected device or keystore cannot sign messages for %s",
+			account.Coin().Code())
+	}
+
+	unused := account.GetUnusedReceiveAddresses()
+	// Use the format hint to get a compatible address
+	if len(format) == 0 {
+		format = defaultScriptType
+	}
+	expectedScriptType, ok := requestAddressScriptTypeMap[format]
+	if !ok {
+		err := fmt.Errorf("Unknown format:  %s", format)
+		return "", "", err
+	}
+	signingConfigIdx := account.Config().Config.SigningConfigurations.FindScriptType(expectedScriptType)
+	if signingConfigIdx == -1 {
+		err := fmt.Errorf("Unknown format: %s", format)
+		return "", "", err
+	}
+	addr := unused[signingConfigIdx].Addresses[0]
+
+	sig, err := keystore.SignBTCMessage(
+		[]byte(message),
+		addr.AbsoluteKeypath(),
+		account.Config().Config.SigningConfigurations[signingConfigIdx].ScriptType(),
+	)
+	if err != nil {
+		return "", "", err
+	}
+
+	return addr.EncodeForHumans(), base64.StdEncoding.EncodeToString(sig), nil
 }
