+++ frontends/web/src/routes/bitsurance/widget.tsx
@@ -18,8 +18,7 @@ import { useTranslation } from 'react-i18next';
 import { useState, useEffect, createRef } from 'react';
 import { RequestAddressV0Message, MessageVersion, parseMessage, serializeMessage, V0MessageType } from 'request-address';
 import { getConfig } from '../../utils/config';
-import { ScriptType } from '../../api/account';
-import { signAddress } from '../../api/exchanges';
+import { ScriptType, signAddress } from '../../api/account';
 import { getInfo } from '../../api/account';
 import { Header } from '../../components/layout';
 import { Spinner } from '../../components/spinner/Spinner';
